+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Version 1.20 (not yet released)
 * Added action to remove a subtype from a contact.
 * Added action to update an existing relationship with an ID.
 * Fixed issue with clearing address fields on Create/Update address.
+* Added action to Find Contact with Role on the Case.
+* Added action to get case data.
+* Added Case ID as a parameter to the create/update activity action.
 
 Version 1.19
 ------------